Scripting: fix package of file moved from backend
 #KT-87922 Fixed
